Add name for actions in GitHubActions generation
