+++ source/Nuke.Common.Tests/CI/ConfigurationGenerationTest.Test_testName=detailed-triggers_attribute=GitHubActionsAttribute.verified.txt
@@ -53,27 +53,30 @@ jobs:
         with:
           submodules: recursive
           fetch-depth: 2
-      - name: Cache .nuke/temp, ~/.nuget/packages
+      - name: 'Cache: .nuke/temp, ~/.nuget/packages'
         uses: actions/cache@v3
         with:
           path: |
             .nuke/temp
             ~/.nuget/packages
           key: ${{ runner.os }}-${{ hashFiles('**/global.json', '**/*.csproj') }}
-      - name: Run './build.cmd Test'
+      - name: 'Run: Test'
         run: ./build.cmd Test
         env:
           OptionalInput: ${{ github.event.inputs.OptionalInput }}
           RequiredInput: ${{ github.event.inputs.RequiredInput }}
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: src'
+        uses: actions/upload-artifact@v3
         with:
           name: src
           path: src
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: test-results'
+        uses: actions/upload-artifact@v3
         with:
           name: test-results
           path: output/test-results
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: coverage-report.zip'
+        uses: actions/upload-artifact@v3
         with:
           name: coverage-report.zip
           path: output/coverage-report.zip
@@ -85,27 +88,30 @@ jobs:
         with:
           submodules: recursive
           fetch-depth: 2
-      - name: Cache .nuke/temp, ~/.nuget/packages
+      - name: 'Cache: .nuke/temp, ~/.nuget/packages'
         uses: actions/cache@v3
         with:
           path: |
             .nuke/temp
             ~/.nuget/packages
           key: ${{ runner.os }}-${{ hashFiles('**/global.json', '**/*.csproj') }}
-      - name: Run './build.cmd Test'
+      - name: 'Run: Test'
         run: ./build.cmd Test
         env:
           OptionalInput: ${{ github.event.inputs.OptionalInput }}
           RequiredInput: ${{ github.event.inputs.RequiredInput }}
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: src'
+        uses: actions/upload-artifact@v3
         with:
           name: src
           path: src
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: test-results'
+        uses: actions/upload-artifact@v3
         with:
           name: test-results
           path: output/test-results
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: coverage-report.zip'
+        uses: actions/upload-artifact@v3
         with:
           name: coverage-report.zip
           path: output/coverage-report.zip
@@ -117,27 +123,30 @@ jobs:
         with:
           submodules: recursive
           fetch-depth: 2
-      - name: Cache .nuke/temp, ~/.nuget/packages
+      - name: 'Cache: .nuke/temp, ~/.nuget/packages'
         uses: actions/cache@v3
         with:
           path: |
             .nuke/temp
             ~/.nuget/packages
           key: ${{ runner.os }}-${{ hashFiles('**/global.json', '**/*.csproj') }}
-      - name: Run './build.cmd Test'
+      - name: 'Run: Test'
         run: ./build.cmd Test
         env:
           OptionalInput: ${{ github.event.inputs.OptionalInput }}
           RequiredInput: ${{ github.event.inputs.RequiredInput }}
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: src'
+        uses: actions/upload-artifact@v3
         with:
           name: src
           path: src
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: test-results'
+        uses: actions/upload-artifact@v3
         with:
           name: test-results
           path: output/test-results
-      - uses: actions/upload-artifact@v3
+      - name: 'Publish: coverage-report.zip'
+        uses: actions/upload-artifact@v3
         with:
           name: coverage-report.zip
           path: output/coverage-report.zip
