+++ source/Nuke.Common/CI/GitHubActions/Configuration/GitHubActionsArtifactStep.cs
@@ -1,0 +1,32 @@
+﻿// Copyright 2023 Maintainers of NUKE.
+// Distributed under the MIT License.
+// https://github.com/nuke-build/nuke/blob/master/LICENSE
+
+using System;
+using System.Linq;
+using Nuke.Common.Utilities;
+
+namespace Nuke.Common.CI.GitHubActions.Configuration
+{
+    public class GitHubActionsArtifactStep : GitHubActionsStep
+    {
+        public string Name { get; set; }
+        public string Path { get; set; }
+
+        public override void Write(CustomFileWriter writer)
+        {
+            writer.WriteLine("- name: " + $"Publish: {Name}".SingleQuote());
+            writer.WriteLine("  uses: actions/upload-artifact@v3");
+
+            using (writer.Indent())
+            {
+                writer.WriteLine("with:");
+                using (writer.Indent())
+                {
+                    writer.WriteLine($"name: {Name}");
+                    writer.WriteLine($"path: {Path}");
+                }
+            }
+        }
+    }
+}
