+++ source/Nuke.Common/CI/GitHubActions/Configuration/GitHubActionsCacheStep.cs
@@ -20,7 +20,7 @@ namespace Nuke.Common.CI.GitHubActions.Configuration
 
         public override void Write(CustomFileWriter writer)
         {
-            writer.WriteLine($"- name: Cache {IncludePatterns.JoinCommaSpace()}");
+            writer.WriteLine("- name: " + $"Cache: {IncludePatterns.JoinCommaSpace()}".SingleQuote());
             using (writer.Indent())
             {
                 writer.WriteLine("uses: actions/cache@v3");
