+++ source/Nuke.Common/CI/GitHubActions/Configuration/GitHubActionsCheckoutStep.cs
@@ -35,25 +35,4 @@ namespace Nuke.Common.CI.GitHubActions.Configuration
             }
         }
     }
-
-    public class GitHubActionsArtifactStep : GitHubActionsStep
-    {
-        public string Name { get; set; }
-        public string Path { get; set; }
-
-        public override void Write(CustomFileWriter writer)
-        {
-            writer.WriteLine("- uses: actions/upload-artifact@v3");
-
-            using (writer.Indent())
-            {
-                writer.WriteLine("with:");
-                using (writer.Indent())
-                {
-                    writer.WriteLine($"name: {Name}");
-                    writer.WriteLine($"path: {Path}");
-                }
-            }
-        }
-    }
 }
